refactor: split data request types up
