+++ contract/src/msgs/data_requests/types/types_tests.rs
@@ -1,6 +1,5 @@
-use test_helpers::{calculate_dr_id_and_args, construct_dr};
+use test::test_helpers::{calculate_dr_id_and_args, construct_dr};
 use testing::MockStorage;
-use types::*;
 
 use super::*;
 use crate::enumerable_status_map;
